universal7580: include: Remove the custom gralloc header
The only use it had was to define some custom gralloc usages
used in the BSP sources. However, those are now defined within
them, so it's no longer necessary.

Having the gralloc header overridden somehow caused some issue
somewhere that causes Codec2 to not work, so with this, Codec2
is now fixed and usable.

Change-Id: I2a41f2812db7874417857de6a212f87e3d7fe782